+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+v8.1.65 (2023-03-09)
+-------------------------
+ * Update endpoints for messages and media
+
+v8.1.64 (2023-03-08)
+-------------------------
+ * Tweak layout for editor
+ * Cleanup fail_old_messages task. Use correct statuses and return number failed.
+
+v8.1.63 (2023-03-08)
+-------------------------
+ * Adjust export download page for new UI
+ * Make media list page (still staff only) filter by org and add index
+
+v8.1.62 (2023-03-08)
+-------------------------
+ * Small z-index tweak
+
+v8.1.61 (2023-03-07)
+-------------------------
+ * Tweak simulator placement in new ui
+
+v8.1.60 (2023-03-07)
+-------------------------
+ * Encourage users to try the new interface
+ * Add lightbox for contact history
+
+v8.1.59 (2023-03-07)
+-------------------------
+ * Rework code depending on msg_type=I|F
+
+v8.1.58 (2023-03-07)
+-------------------------
+ * Add missing channels migration
+ * Use msg.created_by if set in ticket list view
+ * Remove SMS type channel alerts
+
 v8.1.57 (2023-03-06)
 -------------------------
  * Move index on msg.external_id onto the model